+++ app/src/components/root/root.component.editor-tab.schnittstellen.tsx
@@ -28,6 +28,7 @@ import NotInterestedOutlinedIcon from '@mui/icons-material/NotInterestedOutlined
 import Tooltip from '@mui/material/Tooltip';
 import Chip from "@mui/material/Chip";
 import {ElementEditorSectionHeader} from '../element-editor-section-header/element-editor-section-header';
+import {OzgCloudInfo} from '../../modules/destination/components/ozg-cloud-info';
 
 export function RootComponentEditorTabSchnittstellen(props: BaseEditorProps<RootElement, Application>): JSX.Element {
     const api = useApi();
@@ -177,6 +178,27 @@ export function RootComponentEditorTabSchnittstellen(props: BaseEditorProps<Root
                             E-Mail-Konfiguration aufzunehmen. Falls Sie keine E-Mails erhalten, überprüfen Sie bitte auch Ihren Spam-Ordner oder kontaktieren Sie Ihre
                             E-Mail-Administrator:in.
                         </Typography>
+                    </AlertComponent>
+                }
+
+                {
+                    destinations != null &&
+                    destinations.length > 0 &&
+                    props.entity.destinationId != null &&
+                    destinations.find((dest) => dest.id === props.entity.destinationId)?.type === DestinationType.OZGCloud &&
+                    <AlertComponent
+                        title="Hinweis zur OZG-Cloud Schnittstelle"
+                        color="warning"
+                        sx={{
+                            mt: 1,
+                        }}
+                    >
+                        <OzgCloudInfo
+                            sx={{
+                                mb: 0,
+                                mt: 0,
+                            }}
+                        />
                     </AlertComponent>
                 }
             </Box>