+++ src/main/java/de/aivot/GoverBackend/ozgCloud/enums/OZGCloudPostfachAdresseType.java
@@ -1,0 +1,5 @@
+package de.aivot.GoverBackend.ozgCloud.enums;
+
+public enum OZGCloudPostfachAdresseType {
+    Citizen,
+}